Fix for macOS compilation
This is a first attempt at a fix.

The library doesn't compile correctly under macOS
because of two reasons:
1. The native `libtool` provided by Apple doesn't understand the
`--version` flag, the correct command is `libtool -V`. Apple does **not**
provide `libtoolize`.
2. If these tools have been installed by `Homebrew`, the names of
the commands have a _g_ prepended so one has to invoke `glibtool`
and `glibtoolize`.

Again, since `libtoolize` is not provided by Apple, I think it is
safe to assume that both `libtool` and `libtoolize` are installed
by `Homebrew` under macOS.

Anyways, the correct name of the program is parametrized so it can
be supplied via the environment variables `LIBTOOL` and
`LIBTOOLIZE`.